cccd_reader_project/requirements.txt
flask==2.3.3
pillow==10.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
